requirements.txt
fastapi
uvicorn
botbuilder-core